Move finance breakdown to financial section
branch:
