+++ src/components/ResultTable.tsx
@@ -2,7 +2,7 @@ import { Col, Divider, Row } from 'antd'
 import React from 'react'
 import { useTranslation } from 'react-i18next'
 import { ResultData } from '../services/CalculationService'
-import { formatDigits, formatNumber, formatRupiah } from './Formatters'
+import { formatDigits, formatNumber } from './Formatters'
 import { CalculatorSettings, OptimizationTarget } from '../constants'
 import { Panel, renderPanel } from './SolarPanel'
 import { InfoCircleOutlined } from '@ant-design/icons'
@@ -14,12 +14,6 @@ export interface ResultTableProps {
   onOpenDocumentation:  (d: Documentation, title: string) => void
 }
 
-interface BreakEvenPoint {
-  years: number
-  months: number
-}
-
-
 export const ResultTable: React.FunctionComponent<ResultTableProps> = (props) => {
   const { t, i18n } = useTranslation()
   const { results, calculatorSettings, onOpenDocumentation }  = props
@@ -29,12 +23,6 @@ export const ResultTable: React.FunctionComponent<ResultTableProps> = (props) =>
 
   const panels: Panel[] = Array.from(Array(results?.numberOfPanels).keys())
     .map(x =>{ return { index: (x + 1), panelType: x >= results.numberOfPanelsFinancial ? OptimizationTarget.Green: OptimizationTarget.Money } } )
-
-  const monthsInYear = 12
-  const breakEven: BreakEvenPoint = {
-    years: Math.floor(results.breakEvenPointInMonths / monthsInYear),
-    months: Math.round(results.breakEvenPointInMonths % monthsInYear)
-  }
 
   return (
     <div className="ant-table">
@@ -72,37 +60,6 @@ export const ResultTable: React.FunctionComponent<ResultTableProps> = (props) =>
       <Row gutter={12}>
         <Col span={15}>{t('resultTable.monthlyProduction')}</Col>
         <Col span={9}>{`${formatNumber(results.productionPerMonthInKwh, i18n.language)} kWh`}</Col>
-      </Row>
-      <Divider orientation="left">{t('resultTable.financialHeading')}</Divider>
-      <Row gutter={12}>
-        <Col span={15}>{t('resultTable.currentMonthlyCosts')}</Col>
-        <Col span={9}>{formatRupiah(results.currentMonthlyCosts)}</Col>
-      </Row>
-      <Row gutter={12}>
-        <Col span={15}>{t('resultTable.remainingMonthlyCosts')}&nbsp;
-          <InfoCircleOutlined onClick={() => onOpenDocumentation(Documentation.MinimalPayment,t('resultTable.remainingMonthlyCosts'))}/>
-        </Col>
-        <Col span={9}>- {formatRupiah(results.remainingMonthlyCosts)}</Col>
-      </Row>
-      <Row gutter={12}>
-        <Col span={15}>{t('resultTable.monthlyProfit')}</Col>
-        <Col span={9} className='total'>{formatRupiah(results.monthlyProfit)}</Col>
-      </Row>
-      <Row gutter={12}>
-        <Col span={24} >&nbsp;</Col>
-      </Row>
-      <Row gutter={12}>
-        <Col span={15}>{t('resultTable.yearlyProfit')}</Col>
-        <Col span={9}>{formatRupiah(results.yearlyProfit)}</Col>
-      </Row>
-      <Row gutter={12}>
-        <Col span={11}>{t('resultTable.breakEven')}&nbsp;
-          <InfoCircleOutlined onClick={() => onOpenDocumentation(Documentation.RoiExplanation,t('resultTable.breakEven'))}/>
-        </Col>
-        <Col span={13}>{t('resultTable.breakEvenExplanation', { breakEven })}</Col>
-      </Row>
-      <Row gutter={12}>
-        <Col span={24} />
       </Row>
     </div>
   )
